+++ pel216/week6/inc/integration/TrapezoidalRule.h
@@ -29,7 +29,7 @@ namespace pel216 {
 			/**
 			 * @see pel216::week6::IntegrationRule#evaluate()
 			 */
-			virtual double doEvaluate(Function *function, double a, double b, size_t intervals) {
+			virtual double doEvaluate(Function *function, double a, double b, double intervals) {
 
 				double result = 0.0f;
 
@@ -37,7 +37,7 @@ namespace pel216 {
 
 				double a_prime = a;
 				double b_prime = a_prime + alpha;
-				for (size_t idx = 0; idx < intervals; idx++) {
+				for (double idx = 0.0f; idx < intervals; idx++) {
 					result += alpha * ((function->evaluate(a_prime) + function->evaluate(b_prime)) / 2.0f);
 					a_prime += alpha;
 					b_prime += alpha;
